+++ fe/fe-core/src/main/java/org/apache/doris/catalog/FunctionRegistry.java
@@ -20,23 +20,10 @@ package org.apache.doris.catalog;
 import org.apache.doris.nereids.annotation.Developing;
 import org.apache.doris.nereids.exceptions.AnalysisException;
 import org.apache.doris.nereids.trees.expressions.Expression;
-import org.apache.doris.nereids.trees.expressions.functions.AggregateFunction;
-import org.apache.doris.nereids.trees.expressions.functions.Avg;
-import org.apache.doris.nereids.trees.expressions.functions.BoundFunction;
-import org.apache.doris.nereids.trees.expressions.functions.Count;
 import org.apache.doris.nereids.trees.expressions.functions.FunctionBuilder;
-import org.apache.doris.nereids.trees.expressions.functions.Max;
-import org.apache.doris.nereids.trees.expressions.functions.Min;
-import org.apache.doris.nereids.trees.expressions.functions.ScalarFunction;
-import org.apache.doris.nereids.trees.expressions.functions.Substring;
-import org.apache.doris.nereids.trees.expressions.functions.Sum;
-import org.apache.doris.nereids.trees.expressions.functions.WeekOfYear;
-import org.apache.doris.nereids.trees.expressions.functions.Year;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableList;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -51,26 +38,11 @@ import javax.annotation.concurrent.ThreadSafe;
 @Developing
 @ThreadSafe
 public class FunctionRegistry {
-    public static final List<ScalarFunc> SCALAR_FUNCTIONS = ImmutableList.of(
-            scalar(Substring.class, "substr", "substring"),
-            scalar(WeekOfYear.class),
-            scalar(Year.class)
-    );
-
-    public static final ImmutableList<AggregateFunc> AGGREGATE_FUNCTIONS = ImmutableList.of(
-            agg(Avg.class),
-            agg(Count.class),
-            agg(Max.class),
-            agg(Min.class),
-            agg(Sum.class)
-    );
-
     private final Map<String, List<FunctionBuilder>> name2Builders;
 
     public FunctionRegistry() {
         name2Builders = new ConcurrentHashMap<>();
-        addFunctions(name2Builders, SCALAR_FUNCTIONS);
-        addFunctions(name2Builders, AGGREGATE_FUNCTIONS);
+        addBuiltinFunctions(name2Builders);
         afterRegisterBuiltinFunctions(name2Builders);
     }
 
@@ -106,71 +78,15 @@ public class FunctionRegistry {
         return candidateBuilders.get(0);
     }
 
-    private void addFunctions(Map<String, List<FunctionBuilder>> name2FuncBuilders,
-            List<? extends NamedFunc<? extends BoundFunction>> funcs) {
-        for (NamedFunc<? extends BoundFunction> func : funcs) {
-            for (String name : func.names) {
-                if (name2FuncBuilders.containsKey(name)) {
-                    throw new IllegalStateException("Function '" + name + "' already exists in function registry");
-                }
-
-                name2FuncBuilders.put(name, func.functionBuilders);
-            }
-        }
+    private void addBuiltinFunctions(Map<String, List<FunctionBuilder>> name2Builders) {
+        BuiltinFunctions builtinFunctions = new BuiltinFunctions();
+        FunctionHelper.addFunctions(name2Builders, builtinFunctions.scalarFunctions);
+        FunctionHelper.addFunctions(name2Builders, builtinFunctions.aggregateFunctions);
     }
 
     public String getCandidateHint(String name, List<FunctionBuilder> candidateBuilders) {
         return candidateBuilders.stream()
                 .map(builder -> name + builder.toString())
                 .collect(Collectors.joining(", "));
-    }
-
-    public static final ScalarFunc scalar(Class<? extends ScalarFunction> functionClass) {
-        String functionName = functionClass.getSimpleName();
-        return scalar(functionClass, functionName);
-    }
-
-    public static final ScalarFunc scalar(Class<? extends ScalarFunction> functionClass, String... functionNames) {
-        return new ScalarFunc(functionClass, functionNames);
-    }
-
-    public static final AggregateFunc agg(Class<? extends AggregateFunction> functionClass) {
-        String functionName = functionClass.getSimpleName();
-        return new AggregateFunc(functionClass, functionName);
-    }
-
-    private static final AggregateFunc agg(Class<? extends AggregateFunction> functionClass, String... functionNames) {
-        return new AggregateFunc(functionClass, functionNames);
-    }
-
-    /**
-     * use this class to prevent the wrong type from being registered, and support multi function names
-     * like substring and substr.
-     */
-    public static class NamedFunc<T extends BoundFunction> {
-        public final List<String> names;
-        public final Class<? extends T> functionClass;
-
-        public final List<FunctionBuilder> functionBuilders;
-
-        public NamedFunc(Class<? extends T> functionClass, String... names) {
-            this.functionClass = functionClass;
-            this.names = Arrays.stream(names)
-                    .map(String::toLowerCase)
-                    .collect(ImmutableList.toImmutableList());
-            this.functionBuilders = FunctionBuilder.resolve(functionClass);
-        }
-    }
-
-    public static class ScalarFunc extends NamedFunc<ScalarFunction> {
-        public ScalarFunc(Class<? extends ScalarFunction> functionClass, String... names) {
-            super(functionClass, names);
-        }
-    }
-
-    public static class AggregateFunc extends NamedFunc<AggregateFunction> {
-        public AggregateFunc(Class<? extends AggregateFunction> functionClass, String... names) {
-            super(functionClass, names);
-        }
     }
 }
